Move AuthService to authentication module
