+++ app/app.component.ts
@@ -7,7 +7,7 @@ import { IntervalObservable } from 'rxjs/observable/IntervalObservable';
 
 import 'hammerjs';
 
-import { AuthService } from './core/auth/auth.service';
+import { AuthService } from './authentication/auth.service';
 import { NotificationService } from './notifications/notification.service';
 import { ScrollService } from './shared/scroll/scroll.service';
 import { User } from './users/user.model';
