+++ app/app.routing.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule }   from '@angular/router';
 
 import { AuthGuard } from './core/guards/auth-guard';
 import { AuthApiService } from './core/api/auth-api.service';
-import { AuthService } from './core/auth/auth.service';
+import { AuthService } from './authentication/auth.service';
 import { CmsApiService } from './core/api/cms-api.service';
 import { SupervisorGuard } from './core/guards/supervisor-guard';
 import { UserService } from './users/user.service';
