+++ app/authentication/auth.service.ts
@@ -3,10 +3,10 @@ import { Headers } from '@angular/http';
 import { Router } from '@angular/router';
 import { tokenNotExpired, JwtHelper } from 'angular2-jwt';
 
-import { AppComponent } from '../../app.component';
-import { AuthApiService } from '../api/auth-api.service';
-import { ConfigService } from '../../config.service';
-import { UserService } from '../../users/user.service';
+import { AppComponent } from '../app.component';
+import { AuthApiService } from '../core/api/auth-api.service';
+import { ConfigService } from '../config.service';
+import { UserService } from '../users/user.service';
 
 @Injectable()
 export class AuthService {
