+++ app/authentication/authentication.module.ts
@@ -3,9 +3,11 @@ import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { MaterialModule } from '@angular/material';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
+import { AuthHttp, provideAuth } from 'angular2-jwt';
 import { TranslateModule } from 'ng2-translate';
 
 import { authRouting } from './authentication.routing';
+import { AuthService } from './auth.service';
 import { LoginComponent } from './login/login.component';
 import { SignupComponent } from './signup/signup.component';
 
@@ -22,6 +24,19 @@ import { SignupComponent } from './signup/signup.component';
   declarations: [
     LoginComponent,
     SignupComponent
+  ],
+  providers: [
+    AuthService,
+    AuthHttp,
+    provideAuth({
+      headerName: 'Authorization',
+      headerPrefix: 'Bearer',
+      tokenName: 'id_token',
+      tokenGetter: (() => localStorage.getItem('id_token')),
+      globalHeaders: [{'Content-Type': 'application/json'}],
+      noJwtError: true,
+      noTokenScheme: true
+    })
   ]
 })
 export class AuthenticationModule {}
