+++ app/authentication/signup/signup.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
-import { AuthService } from '../../core/auth/auth.service';
+import { AuthService } from '../auth.service';
 import { User } from '../../users/user.model';
 
 @Component({
