+++ app/core/core.module.ts
@@ -1,7 +1,6 @@
 import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { HttpModule } from '@angular/http';
 import { RouterModule } from '@angular/router';
-import { AuthHttp, provideAuth } from 'angular2-jwt';
 import { ToasterModule, ToasterService } from 'angular2-toaster';
 import { NgxPaginationModule } from 'ngx-pagination';
 import { TranslateModule, TranslateService } from 'ng2-translate';
@@ -9,7 +8,6 @@ import { TranslateModule, TranslateService } from 'ng2-translate';
 import { AdminGuard } from './guards/admin-guard';
 import { AuthApiService } from './api/auth-api.service';
 import { AuthGuard } from './guards/auth-guard';
-import { AuthService } from './auth/auth.service';
 import { CmsApiService } from './api/cms-api.service';
 import { OOApiService } from './api/oo-api.service';
 import { SupervisorGuard } from './guards/supervisor-guard';
@@ -30,23 +28,12 @@ import { SupervisorGuard } from './guards/supervisor-guard';
   providers: [
     TranslateService,
     ToasterService,
-    AuthService,
     AuthApiService,
     CmsApiService,
     OOApiService,
     AuthGuard,
     AdminGuard,
-    SupervisorGuard,
-    AuthHttp,
-    provideAuth({
-      headerName: 'Authorization',
-      headerPrefix: 'Bearer',
-      tokenName: 'id_token',
-      tokenGetter: (() => localStorage.getItem('id_token')),
-      globalHeaders: [{'Content-Type': 'application/json'}],
-      noJwtError: true,
-      noTokenScheme: true
-    })
+    SupervisorGuard
   ]
 })
 export class CoreModule {
